app/requirements.txt
Flask==3.0.3
Pillow==10.4.0
qrcode[pil]==7.4.2
# Pin numpy <2 — LivePortrait and torch 2.3 require numpy 1.x.
numpy>=1.24,<2
opencv-python>=4.9,<4.11
imageio>=2.34
imageio-ffmpeg>=0.5.1
tqdm>=4.66
requests>=2.31
